+++ scripts/migration-tests.sh
@@ -1,0 +1,1 @@
+BUILD_DUMMY_WASM_BINARY=1 cargo test -p migration-tests --lib -- --nocapture
